+++ test/brando/utils_test.exs
@@ -35,11 +35,11 @@ defmodule Brando.UtilsTest do
 
   test "slugify special characters" do
     assert slugify("special characters (#?@$%^*) are also ASCII")
-           == "special-characters-at-are-also-ascii"
+           == "special-characters-atdollar-are-also-ascii"
   end
 
   test "slugify & -> and" do
-    assert slugify("tom & jerry") == "tom-jerry"
+    assert slugify("tom & jerry") == "tom-and-jerry"
   end
 
   test "slugify strip extraneous dashes" do
@@ -50,7 +50,7 @@ defmodule Brando.UtilsTest do
     assert slugify_filename("testing with spaces.jpeg")
            == "testing-with-spaces.jpeg"
     assert slugify_filename("-start æøå-.jpeg")
-           == "start-aeoa.jpeg"
+           == "start-aeoeaa.jpeg"
   end
 
   test "random_filename/1" do
@@ -107,7 +107,10 @@ defmodule Brando.UtilsTest do
   end
 
   test "img_url/2" do
-    img = %{sizes: %{"thumb" => "images/thumb/file.jpg"}}
+    img = %{
+      path: "original/path/file.jpg",
+      sizes: %{"thumb" => "images/thumb/file.jpg"}
+    }
     assert img_url(img, :thumb)
            == "images/thumb/file.jpg"
     assert img_url(nil, :thumb, [default: "default.jpg", prefix: "prefix"])
@@ -120,6 +123,12 @@ defmodule Brando.UtilsTest do
            == "images/thumb/file.jpg"
     assert img_url(img, "thumb", [default: "default.jpg"])
            == "images/thumb/file.jpg"
+    assert img_url(img, :original)
+           == "original/path/file.jpg"
+    assert img_url(img, :original, prefix: "prefix")
+           == "prefix/original/path/file.jpg"
+
+
     assert_raise ArgumentError, fn ->
       img_url(img, :notasize, [default: "default.jpg"])
     end